generalize vote model

+++ dag/dag.go
@@ -1054,7 +1054,7 @@ func (dag *Dag) GetElectedMediatorsAddress() ([]common.Address, error) {
 		return nil, err
 	}
 	MediatorNumber := gp.GetActiveMediatorCount()
-	return dag.statedb.GetSortedVote(uint(MediatorNumber), 0, 0)
+	return dag.statedb.GetSortedVote(uint8(MediatorNumber), 0, 0)
 }
 
 // UpdateMediator
